Rocketseat - Nível 4 - Desafio Database Relations - Create Customer

+++ src/shared/infra/typeorm/migrations/1599596688798-CreateOrderProducts.ts
@@ -22,7 +22,7 @@ export default class CreateOrderProducts1599596688798
                     },
                     {
                         name: 'quantity',
-                        type: 'decimal',
+                        type: 'int',
                     },
                     {
                         name: 'created_at',
